update: update on ianni sell margin and help modal size
update on the margin of iannistore, from 1.7 * to 2.154. Fix on the help modal size, coming from a small modal to a large modal, becaming more apleased to see with large screens. Was removed as well the store pictures on the help modal, because it was self explanatory on the table